refactor(421): extract WASM heap-marshalling helpers to vera/runtime/heap.py
Step 3 of the codegen/api.py decomposition (#421): the ~820-line heap-marshalling helper layer -- memory read/write, GC shadow-stack rooting (_ShadowGuard), wrapper-ADT handle tagging (_wrap_handle), the Map/Set bucket codec, and Result/Option/Array allocation -- lifts out of execute() into vera/runtime/heap.py. The helpers were already parameterised by the wasmtime.Caller, so they become plain module-level functions; the one closed-over 'store' (_read_wasm_string) is fixed to use 'caller' like its siblings.

execute() and the still-nested host families call them via a module-level import (35 names; 17 stay heap-internal). The dead _validate_wrap_handle import is dropped from api.py. Repointed the source-location test test_read_wasm_string_uses_errors_replace at heap.py (the function moved). api.py 3,790->3,001 lines. mypy + ruff clean; the full suite (4,726), 92 conformance, 35 examples, and the #734 gate stay green.
